research-apparatus/backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
gunicorn==21.2.0
pydantic-settings==2.0.3
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
numpy==1.26.1
pandas==2.1.1